+++ astropy/io/votable/exceptions.py
@@ -1448,6 +1448,14 @@ class E23(VOTableSpecWarning):
     default_args = ('x',)
 
 
+class E25(VOTableSpecWarning):
+    """
+    A VOTable cannot have a DATA section without any defined FIELD; DATA will be ignored.
+    """
+
+    message_template = "No FIELDs are defined; DATA section will be ignored."
+
+
 def _get_warning_and_exception_classes(prefix):
     classes = []
     for key, val in globals().items():